git-svn-id: http://gforge.hl7.org/svn/fhir/trunk@1265 2f0db536-2c49-4257-a3fa-e771ed206c19

+++ tools/java/org.hl7.fhir.tools.core/src/org/hl7/fhir/tools/publisher/BookMaker.java
@@ -123,7 +123,8 @@ public class BookMaker {
 						  found = findTarget(pages.get(parts[0]), parts[1]);
 						  if (!found)
 							try {
-								new XhtmlComposer().compose(new FileOutputStream("c:\\temp\\source.htm"), pages.get(parts[0]));
+								if (new File("c:\\temp\\source.htm").exists())
+								  new XhtmlComposer().compose(new FileOutputStream("c:\\temp\\source.htm"), pages.get(parts[0]));
 							} catch (Exception e) {
 								e.printStackTrace();
 							}
